// packages/token-lab/src/styles/tokens.d.ts
export const border: {
  "1": string,
  "2": string,
  "3": string,
  "focus": string
};
export const color: {
  "base": {
    "blue": {
      "100": string,
      "200": string,
      "300": string,
      "400": string,
      "500": string,
      "600": string,
      "700": string,
      "800": string,
      "900": string,
      "1000": string
    },
    "gray": {
      "100": string,
      "200": string,
      "300": string,
      "400": string,
      "500": string,
      "600": string,
      "700": string,
      "800": string,
      "900": string,
      "950": string,
      "1000": string
    },
    "lime": {
      "100": string,
      "200": string,
      "300": string,
      "400": string,
      "500": string,
      "600": string,
      "700": string,
      "800": string,
      "900": string,
      "1000": string
    },
    "orange": {
      "100": string,
      "200": string,
      "300": string,
      "400": string,
      "500": string,
      "600": string,
      "700": string,
      "800": string,
      "900": string,
      "1000": string
    }
  },
  "bg": {
    "1": string,
    "2": string,
    "3": string,
    "error": string,
    "selected": string,
    "selectedHint": string
  },
  "border": {
    "1": string,
    "2": string,
    "3": string,
    "focus": string,
    "selected": string
  },
  "icon": {
    "100": string,
    "200": string,
    "selected": string
  },
  "text": {
    "1": string,
    "2": string,
    "3": string,
    "error": string,
    "selected": string
  }
};
export const ease: {
  "linear": string,
  "quintIn": string,
  "quintInOut": string,
  "quintOut": string
};
export const font: {
  "body": {
    "fontFamily": string,
    "fontVariationSettings": string,
    "fontWeight": string,
    "fontSize": string,
    "letterSpacing": string,
    "lineHeight": string
  },
  "bodyLongform": {
    "fontFamily": string,
    "fontVariationSettings": string,
    "fontWeight": string,
    "fontSize": string,
    "letterSpacing": string,
    "lineHeight": string
  },
  "bodyStrong": {
    "fontFamily": string,
    "fontVariationSettings": string,
    "fontWeight": string,
    "fontSize": string,
    "letterSpacing": string,
    "lineHeight": string
  },
  "heading1": {
    "fontFamily": string,
    "fontVariationSettings": string,
    "fontWeight": string,
    "fontSize": string,
    "lineHeight": string,
    "letterSpacing": string
  },
  "heading2": {
    "fontFamily": string,
    "fontVariationSettings": string,
    "fontWeight": string,
    "fontSize": string,
    "lineHeight": string,
    "letterSpacing": string
  },
  "heading3": {
    "fontFamily": string,
    "fontVariationSettings": string,
    "fontWeight": string,
    "fontSize": string,
    "lineHeight": string,
    "letterSpacing": string
  },
  "heading4": {
    "fontFamily": string,
    "fontVariationSettings": string,
    "fontWeight": string,
    "fontSize": string,
    "letterSpacing": string,
    "lineHeight": string
  },
  "heading5": {
    "fontFamily": string,
    "fontVariationSettings": string,
    "fontWeight": string,
    "fontSize": string,
    "letterSpacing": string,
    "lineHeight": string
  },
  "heading6": {
    "fontFamily": string,
    "fontVariationSettings": string,
    "fontWeight": string,
    "fontSize": string,
    "letterSpacing": string,
    "lineHeight": string
  },
  "labelSmall": {
    "fontFamily": string,
    "fontVariationSettings": string,
    "fontWeight": string,
    "fontSize": string,
    "lineHeight": string,
    "letterSpacing": string
  },
  "mono": string,
  "sans": string
};
export const layer: {
  "base": string,
  "menu": string,
  "modal": string,
  "nav": string,
  "popover": string,
  "tooltip": string
};
export const radius: {
  "100": string,
  "200": string,
  "300": string,
  "400": string,
  "500": string,
  "1000": string,
  "000": string
};
export const shadow: {
  "100": string,
  "300": string
};
export const space: {
  "100": string,
  "200": string,
  "300": string,
  "400": string,
  "500": string,
  "600": string,
  "700": string,
  "800": string,
  "900": string,
  "000": string,
  "controlL": string,
  "controlM": string,
  "controlS": string
};